refactor(browser): isolate native overlay reconciliation
