+++ src/components/browser-pane.tsx
@@ -12,6 +12,14 @@ import {
   type TauriBrowserBridge,
 } from "@/lib/browser-native-bridge";
 import {
+  BROWSER_MOTION_WINDOW_MS,
+  BROWSER_RECONCILE_INTERVAL_MS,
+  nodeContainsNativeWebviewCover,
+  recordBrowserReconcile,
+  surfaceIsCovered,
+  WEBVIEW_OFFSCREEN,
+} from "@/lib/browser-native-overlay";
+import {
   createExpectedBrowserNavigation,
   decideBrowserNavigationEvent,
   type BrowserNavigationRequest,
@@ -57,68 +65,6 @@ const NATIVE_BROWSER_LABEL_PREFIX = "cave-browser-";
 //      (drag-to-split drop targets, custom covers) sits over it. Transient
 //      live regions (toasts) are ignored so a corner toast doesn't blank the
 //      page.
-function surfaceIsCovered(surface: HTMLElement, rect: DOMRect): boolean {
-  const overlays = document.querySelectorAll(NATIVE_WEBVIEW_COVER_SELECTOR);
-  for (const overlay of overlays) {
-    if (overlay.getClientRects().length > 0) return true;
-  }
-  const inset = 12;
-  const points: Array<[number, number]> = [
-    [rect.left + rect.width / 2, rect.top + rect.height / 2],
-    [rect.left + inset, rect.top + inset],
-    [rect.right - inset, rect.top + inset],
-    [rect.left + inset, rect.bottom - inset],
-    [rect.right - inset, rect.bottom - inset],
-  ];
-  for (const [x, y] of points) {
-    const hit = document.elementFromPoint(x, y);
-    if (!hit || surface.contains(hit)) continue;
-    if (hit.closest('[role="status"], [role="alert"], [aria-live]')) continue;
-    return true;
-  }
-  return false;
-}
-
-const NATIVE_WEBVIEW_COVER_SELECTOR =
-  '[role="dialog"], [aria-modal="true"], [role="menu"], [role="listbox"], [data-native-webview-cover="true"]';
-const BROWSER_RECONCILE_INTERVAL_MS = 100;
-const BROWSER_MOTION_WINDOW_MS = 400;
-const BROWSER_RECONCILE_METRICS_KEY = "__CAVE_BROWSER_RECONCILE_METRICS__";
-
-type BrowserReconcileMetrics = {
-  count: number;
-  totalDurationMs: number;
-  lastDurationMs: number;
-  startedAt: number;
-};
-
-function recordBrowserReconcile(durationMs: number): void {
-  const metricsWindow = window as typeof window & {
-    [BROWSER_RECONCILE_METRICS_KEY]?: BrowserReconcileMetrics;
-  };
-  const metrics = metricsWindow[BROWSER_RECONCILE_METRICS_KEY] ?? {
-    count: 0,
-    totalDurationMs: 0,
-    lastDurationMs: 0,
-    startedAt: Date.now(),
-  };
-  metrics.count += 1;
-  metrics.totalDurationMs += durationMs;
-  metrics.lastDurationMs = durationMs;
-  metricsWindow[BROWSER_RECONCILE_METRICS_KEY] = metrics;
-}
-
-function nodeContainsNativeWebviewCover(node: Node): boolean {
-  return node instanceof Element && (
-    node.matches(NATIVE_WEBVIEW_COVER_SELECTOR) ||
-    node.querySelector(NATIVE_WEBVIEW_COVER_SELECTOR) !== null
-  );
-}
-
-// Mirrors OFFSCREEN_X/OFFSCREEN_Y in src-tauri/src/browser.rs — the "hidden"
-// position for a native webview.
-const WEBVIEW_OFFSCREEN = -10000;
-
 export type BrowserPaneHandle = {
   navigateTo: (url: string) => void;
 };
